ensured only deleted entities are not upserted first

+++ service/sap-hana-service.py
@@ -34,12 +34,14 @@ def get_rows(schemaname,tablename):
             user=HANA_USER,
             password=HANA_PASS
         )
-        logger.info("Connected to Hana, running query {}".format(query))
+        logger.info("Connected to Hana")
     except:
         logger.info("Couldn't connect to Hana")
         return Response(status=403)
 
     def emit_rows(connection,querystring):
+        logger.info("Running query {}".format(querystring))
+
         yield "["
 
         ### rest of your code here ###
@@ -116,10 +118,12 @@ def put_rows(schemaname,tablename):
     row_data = ()
     deleted_ids = {}
     for entity in entities:
-        row_temp = ()
-        for column in columns:
-            row_temp = row_temp + (entity[column],)
-        row_data = row_data + (row_temp,)
+        # add to data to insert if not a delete
+        if(entity['_deleted']==False):
+            row_temp = ()
+            for column in columns:
+                row_temp = row_temp + (entity[column],)
+            row_data = row_data + (row_temp,)
         # stores the last known state of the _deleted property - so we can delete after upserts are completed.
         temp_delete_dict = {}
         temp_delete_dict['_deleted'] = entity['_deleted']
@@ -135,34 +139,40 @@ def put_rows(schemaname,tablename):
                 delete_temp = delete_temp + (deleted_ids[deleted][key],)
             delete_data = delete_data + (delete_temp,)
 
-    # Set up a parameterized SQL INSERT
-    parms = ("?," * len(row_data[0]))[:-1]
-    query = "UPSERT " + schemaname + "." + tablename + " " + columns_text + " VALUES (%s) WITH PRIMARY KEY;" % (parms)
-
-    # Set up parameterized SQL DELETE keys in WHERE conditional
-    key_conditional = ''
-    isFirst = True
-    for key_col in table_keys:
-        if(isFirst):
-            isFirst = False
-        else:
-            key_conditional = key_conditional + "AND "
-        key_conditional = key_conditional + key_col + " = ? "
-
-    delete_query = "DELETE FROM " + schemaname + "." + tablename + " WHERE " + key_conditional
-
     try:
-        ## upsert rows into hana
-        logger.info("Running upserts: " + query)
-        cursor = conn.cursor()
-        cursor.executemany(query, row_data)
-        cursor.close()
+        # Set up a parameterized SQL 
+        if(len(row_data)!=0):
+            parms = ("?," * len(row_data[0]))[:-1]
+            query = "UPSERT " + schemaname + "." + tablename + " " + columns_text + " VALUES (%s) WITH PRIMARY KEY;" % (parms)
+    
+            ## upsert rows into hana
+            logger.info("Running " + "{}".format(len(row_data)) + " upserts: " + query)
+            cursor = conn.cursor()
+            cursor.executemany(query, row_data)
+            cursor.close()
+        else:
+            logger.info("No new rows")
 
-        ## upsert rows into hana
-        logger.info("Running deletes: " + delete_query)
-        cursor_del = conn.cursor()
-        cursor_del.executemany(delete_query, delete_data)
-        cursor_del.close()
+        # Set up parameterized SQL DELETE keys in WHERE conditional
+        if(len(delete_data)!=0):
+            key_conditional = ''
+            isFirst = True
+            for key_col in table_keys:
+                if(isFirst):
+                    isFirst = False
+                else:
+                    key_conditional = key_conditional + "AND "
+                key_conditional = key_conditional + key_col + " = ? "
+        
+            delete_query = "DELETE FROM " + schemaname + "." + tablename + " WHERE " + key_conditional
+
+            ## upsert rows into hana
+            logger.info("Running " + "{}".format(len(delete_data)) + " deletes: " + delete_query)
+            cursor_del = conn.cursor()
+            cursor_del.executemany(delete_query, delete_data)
+            cursor_del.close()
+        else:
+            logger.info("No new deletes")
 
         return Response(status=200)
